Added better colors and info messages

+++ clean.sh
@@ -1,12 +1,21 @@
+# Color defines
+errorColor="\033[0;31m"
+noColor="\033[0m"
+infoColor="\033[0;36m"
+successColor="\033[0;32m"
+
 # Cleans your .cache folder
+echo -e "${infoColor}Cleaning caches${noColor}"
 rm -rf ~/.cache/
+
+echo -e "${infoColor}Cleaning ccaches${noColor}"
 rm -rf ~/.ccache/
 
 # Garbage collects all your git repos
 gits="$(find | grep '\.git$')"
 home="$(pwd)"
 while read -r git; do
-    echo "Cleaning $git"
+    echo -e "${infoColor}Cleaning ${successColor}$git${noColor}"
     res="$(cd ${git::-4} && git gc && cd $home)"
 done <<< "$gits"
 
@@ -14,6 +23,11 @@ done <<< "$gits"
 # This can get really intense when you take 162
 classes="$(find | grep '\.class$')"
 while read -r class; do
-    echo "Cleaning $class"
+    if [ -z "$class" ]
+    then
+        echo -e "${infoColor}No classes to clean!${noColor}"
+        break;
+    fi
+    echo -e "${infoColor}Cleaning ${successColor}$class${noColor}"
     res="$(rm $class)"
 done <<< "$classes"
